requirements.txt
pathway==0.8.2
llama-index
requests==2.31.0
python-dotenv==1.0.1
fpdf==1.7.2
pandas
flask